+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Changelog.rst
@@ -9,6 +9,11 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+4.0.3
+-----
+Changed
+*******
+- Use Stormancer.Abstraction.Server 4.2.0.2-pre.
 4.0.2
 -----
 Changed
@@ -19,7 +24,7 @@ Changed
 -----
 Changed
 *******
-- Use Stormancer.Abstraction.Server 6.1.0
+- Use Stormancer.Abstraction.Server 4.1.0
 
 4.0.0.2
 -------