+++ src/Stormancer.Plugins/Profile/Stormancer.Server.Plugins.Profile/Changelog.rst
@@ -9,6 +9,11 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+2.1.1
+-----
+Changed
+*******
+- Use Stormancer.Abstraction.Server 4.2.0.2-pre
 2.1.0.6
 -------
 Added